requirements.txt
fastapi
uvicorn
google-generativeai
sqlalchemy
pymysql
python-dotenv 